requirements.txt
streamlit
numpy
matplotlib
tensorflow==2.20.0
Pillow
